feat: Add'l moves to api-authentication for shared files
