+++ api-authentication/src/main/java/com/synopsys/integration/alert/api/authentication/descriptor/AuthenticationDescriptor.java
@@ -5,16 +5,16 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.component.authentication.descriptor;
+package com.synopsys.integration.alert.api.authentication.descriptor;
 
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import com.synopsys.integration.alert.api.authentication.validator.AuthenticationConfigurationFieldModelValidator;
 import com.synopsys.integration.alert.common.descriptor.ComponentDescriptor;
 import com.synopsys.integration.alert.common.descriptor.validator.GlobalConfigurationFieldModelValidator;
-import com.synopsys.integration.alert.component.authentication.validator.AuthenticationConfigurationFieldModelValidator;
 
 @Component
 public class AuthenticationDescriptor extends ComponentDescriptor {
