+++ api-authentication/src/main/java/com/synopsys/integration/alert/api/authentication/descriptor/AuthenticationDescriptorKey.java
@@ -5,7 +5,8 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.component.authentication.descriptor;
+package com.synopsys.integration.alert.api.authentication.descriptor;
+
 
 import org.springframework.stereotype.Component;
