+++ authentication-ldap/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation project(':alert-database')
     implementation project(':api-common-model')
     implementation project(':api-descriptor')
+    implementation project(':api-authentication')
     implementation project(':component')
 
     implementation 'jakarta.persistence:jakarta.persistence-api'
