+++ authentication-saml/build.gradle
@@ -5,6 +5,7 @@ dependencies {
 
     implementation project(':alert-common')
     implementation project(':alert-database')
+    implementation project(':api-authentication')
     implementation project(':api-common-model')
     implementation project(':api-descriptor')
     implementation project(':api-environment')
