+++ component/src/main/java/com/synopsys/integration/alert/component/authentication/web/SAMLMetadataUploadFunctionController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.synopsys.integration.alert.common.action.upload.AbstractUploadAction;
 import com.synopsys.integration.alert.common.rest.api.AbstractUploadFunctionController;
-import com.synopsys.integration.alert.component.authentication.descriptor.AuthenticationDescriptor;
+import com.synopsys.integration.alert.api.authentication.descriptor.AuthenticationDescriptor;
 
 @Controller
 @RequestMapping(SAMLMetadataUploadFunctionController.SAML_UPLOAD_URL)
